Bug fixes
Fixed replying to a thread with an image directing the user back to the
index instead of back to the thread.
Fixed the direct link on a reply to a thread page directing to an
invalid thread link.
Fixed the gap between buttons on registration page.
Adjusted text for owner-specific view counter.